Add files via upload
I learnt how to build this quiz from this tutorial I found on youtube: https://www.youtube.com/watch?v=4TZ1K8EHT2M&t=565s 
It was a little difficult at first because although the code is very simple to follow and understand, the syntax errors kept on showing up and even when I would compare my code to the tutorial it wasn't obvious until I spent time scanning each and every line and realizing I had missed a comma here or I didn't move the brackets to the correct position. I did enjoy how entertaining the actual code was because it was nice to see the input and output of the code when I interacted with the quiz (when I was answering the questions).