Add files via upload
I found this book called "automatetheboringstuff" and I was using this to practice the basics of python and in this screenshot here I was basically trying to understand how to use operators, floats, integers, and strings in conjunction and see how well I can execute various commands. It was pretty easy to begin with but once your start adding in variables and if statements it starts becoming very difficult because it's not just about linking the variables to the correct if statements its also about making sure you keep your code neat and easy to read otherwise it becomes very difficult to go back and fix where you went wrong.