Add files via upload
I followed a tutorial for this game (https://www.youtube.com/watch?v=4TZ1K8EHT2M&t=565s), and it was pretty fun until I realized I couldn't actually complete the game because the upper case letters weren't being recognized which frustrated me because I knew the code was right but the output was still incorrect. So I had to keep on guessing the letters until I could visually understand the terminal had recognized the lower case (a) but not the upper case (A). I think I also should've added comments explaining how the loop function works in this one because it's something I still struggle to understand on how it works. Also I need to develop a habit of adding in more comments (# statements) that explain what each section of code is doing.