Add files via upload
I found this mind blowing because I come from a design background, and this is something that would take me almost a week to accomplish because combining two different compositions is very difficult, so seeing this not only executed perfectly but also getting to know that this is one of the many possibilities with coding is definitely very intriguing. I want to see how I can possibly use this stye of coding to create short animation clips to make the artwork sort of 'come alive' more. I would love to try this once I'm more competent with python and C++.